--- libs/agno/agno/os/interfaces/playground/__init__.py
@@ -1,3 +1,0 @@
-from agno.os.interfaces.playground.playground import Playground
-
-__all__ = ["Playground"]
